+++ tests/behat/Writer/MyObjectWriter.php
@@ -50,7 +50,7 @@ class MyObjectWriter implements WriterInterface
      */
     public function save(
         ObjectInterface $object,
-        PromiseInterface $promise = null,
+        ?PromiseInterface $promise = null,
         ?bool $preferRealDateOnUpdate = null,
     ): WriterInterface {
         $this->persist($object, $promise, $preferRealDateOnUpdate);